Extend RPC command "searchrawtransactions"
"searchrawtransactions" is still compatible to the original command, but it's documentation was improved, and orphaned transactions can optionally be excluded from the results.

Description:

> searchrawtransactions "address" ( verbose skip count includeorphans )

Returns an array of all confirmed transactions associated with address.

Note: as per default, orphaned transactions, which are not part of the active chain, are included in the results.

Arguments:
1. address          (string, required) The Bitcoin address
2. verbose          (numeric, optional, default=0) If 0, return only transaction hex
3. skip             (numeric, optional, default=0) The number of transactions to skip
4. count            (numeric, optional, default=100) The number of transactions to return
5. includeorphans   (numeric, optional, default=1) If 0, exclude orphaned transactions