Add coins view cache for transaction inputs
An empty dummy backend is used for the view, so it's just a smart container, which held in the memory.

The cache is intended to hold transaction inputs, and it further allows the injection of entries. It's possible to bind the view to the mempool or pcoinstip, and directly fetch unspent outputs (if available), instead of full transactions (read from disk [or maybe cache]).

The coins view cache has a nice interface, and provides handy methods such as GetOutputFor(CTxIn), GetValueIn(CTransaction).

Since it's basically just a wrapper around a map (in memory), using these is fast.

It is required to populate the cache explicitly.